nmpolicy, types: Move the API types to a dedicated package
Internal packages need to use the API types, and the API layer needs to
call the inner sub-packages.

To avoid a dependency loop, the API types have been moved to a separate
package.
