+++ nmpolicy/types/types.go
@@ -13,7 +13,7 @@
 // limitations under the License.
 //
 
-package nmpolicy
+package types
 
 import "time"
